refactor cartoon geometry generation
the main reason for the change is to create a MeshGeom for each
chain (with non-zero trace residues). Putting all geometry in
one chain would be more efficient but we tend to generate too
many vertices to be addressed with an unsigned short.

as a welcome side-effect, the cartoon geometry generation code is
easier to read.