split geometry at chain boundaries
Render geometry is split at boundaries of chains, even if there is more
space available in the vertex array. This change is required for rendering
of molecular assemblies, because symmetry operators are applied on a
per-chain basis.

This commit also introduces two new internal classes LineChainData and
MeshChainData, deriving from VertexArray and IndexedVertexArray to hold
molecule-related attributes and functionality.