add back the center on atom doubleClick behavior
This restores the 'centering' behaviour, but makes it possible
to opt out of it by passing a null atomDoubleClick when constructing
the viewer. addListener now also understands the special callback
value 'center' to install the center on atom event listener, which is
granted a little tricky to write because of all the transformation
code.