add method to get current atom color
It uses the trace/atom assoc to figure out what the color of an atom
currently is. It will have to run through the whole assoc, because
we don't maintain a map from atom to vertex assoc entry. This can
potentially get slow when requesting the color of many atoms. If
that use-case every comes up, we might add a specialized method to
do that which runs over the vert assoc once and builds a map.