improve animation handling
The change was necessary to properly handle user input during spinning/
rock and rolling. Now, the user is still able to rotate/zoom/pan the
camera while spin/rockAndRoll are active.

It's now also possible to have both rockAndRoll and spinning enabled at
the same time without one of the two canceling the effect of the other.

The drawback of this change is that the animation classes are now tied
to the camera. Before, it would have been possible to reuse the same
animation classes to move other objects in the scene.