increase number of pickable objects to 2^24
This is done by also using the alpha channel for the object id. This
needs to be tested on mobile devices as it's possible the precision
for float/int in the shaders are not high enough for this to work
properly. In the worst case picking does not work as expected, but
at least the object gets drawn on the screen.